Move comic-making back into a single file
generateComic.py is now a complete file, rather than a driver shell
makeComic.py has gone the way of the Dodo
